requirements.txt
dotenv
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=1.9.0
numpy>=1.21.0
websockets>=10.0

openai-whisper>=20231117
# or faster-whisper (faster alternative)
# faster-whisper>=0.9.0
